+++ src/test/kotlin/com/cmgapps/license/reporter/XmlReportShould.kt
@@ -1,17 +1,7 @@
 /*
  * Copyright (c) 2019. Christian Grach <christian.grach@cmgapps.com>
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 package com.cmgapps.license.reporter
@@ -31,54 +21,39 @@ class XmlReportShould {
             `is`(
                 """
                     <?xml version="1.0" encoding="UTF-8" ?>
-                    <libraries>
-                      <library>
+                    <libraries xmlns="https://www.cmgapps.com" xsi:schemaLocation="https://www.cmgapps.com https://www.cmgapps.com/xsd/licenses.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+                      <library id="Test_lib_1_1.0" version="1.0">
                         <name>
                           Test lib 1
                         </name>
-                        <version>
-                          1.0
-                        </version>
                         <description>
                           proper description
                         </description>
                         <licenses>
-                          <license>
+                          <license url="http://www.apache.org/licenses/LICENSE-2.0.txt">
                             <name>
                               Apache 2.0
                             </name>
-                            <url>
-                              http://www.apache.org/licenses/LICENSE-2.0.txt
-                            </url>
                           </license>
-                          <license>
+                          <license url="http://opensource.org/licenses/MIT">
                             <name>
                               MIT License
                             </name>
-                            <url>
-                              http://opensource.org/licenses/MIT
-                            </url>
                           </license>
                         </licenses>
                       </library>
-                      <library>
+                      <library id="Test_lib_2_2.3.4" version="2.3.4">
                         <name>
                           Test lib 2
                         </name>
-                        <version>
-                          2.3.4
-                        </version>
                         <description>
                           descriptions of lib 2
                         </description>
                         <licenses>
-                          <license>
+                          <license url="http://www.apache.org/licenses/LICENSE-2.0.txt">
                             <name>
                               Apache 2.0
                             </name>
-                            <url>
-                              http://www.apache.org/licenses/LICENSE-2.0.txt
-                            </url>
                           </license>
                         </licenses>
                       </library>